------------------------------------------------------------------------ r3583 | lunbr01 | 2010-09-09 18:00:15 -0400 (Thu, 09 Sep 2010) | 71 lines
This change resolves a regression introduced by the prior fix
(for bug 124325) to the Windows tcp_ip protocol driver.  The prior
fix introduced a call to DisconnectEx() just before the socket is
closed in GCC_DISCONNECT to ensure a graceful shutdown of the session
(specifically, to ensure that sent messages were received by the
other side of the connection).  This worked quite well for normal
disconnects and prevented the occasional hang caused by a GCA_RELEASE
message not being received by the DBMS (even though it was successfully
"sent").

The problem arises when an Ingres program uses GCA_SAVE/GCA_RESTORE
to share a connection.  As soon as one of the processes sharing the
connection disconnects, the connection became "truly" shutdown and
was unusable by the other process.  For instance, one process in
createdb (call him "parent") would connect to the dbms and do
some work, then call GCA_SAVE to pass the connection to a spawned
process (call him "child").  The child would do work on the
connection and then during termination would issue a disconnect.
The parent would then try to resume processing on the same connection,
but this would fail as the TCP/IP connection was shut down.
When just closing the socket in the child, but not calling
DisconnectEx(), the parent is able to resume processing on the
shared socket.  So, the fix is to not issue the DisconnectEx()
if a socket is shared.  A flag was added to the PCB to keep track
of this and is set whenever GCwinsock2_save() or GCwinsock2_restore()
are called.  Ie, the socket is marked as shared in both the parent
and the child.  It is not sufficient to just mark it in the parent
or the child because there is no guarantee what order the processes
will disconnect from the session.  With Ingres gateways, the parent
(gateway daemon) disconnects right away after spawning the child.
It may be desirable in the future to keep a use count for a
connection so that only the last disconnecting process issues
the DisconnectEx().  However, this would require some sort of
cross-process communication object/counter (like a mutex, semaphore,
shared memory, or setting some option on the socket--kind of a hack,
but linger timeout value would likely work since linger on Windows
with async/overlapped sockets doesn't seem to work anyway so is
unused by the driver).  With the current fix, once a socket is
marked "shared", no DisconnectEx() will be issued.  Hence, it
is theoretically possible that a shared session might not shut
down gracefully after the last disconnect.  However, this has never
been seen in any testing; the problem with lost sent GCA_RELEASE
message has only been seen (occasionally at that) in the GCC to DBMS
local connection and GCC does not use GCA_SAVE/GCA_RESTORE.

Network disconnects (named pipes and TCP/IP) have always been
problematic with Ingres.  So, I am documenting here, as a sidenote,
some other options I researched and, in some cases tested, as an
alternative to DisconnectEx() for ensuring a graceful shutdown (no
lost sent messages) but still allow processes to share the session,
even after one process disconnects....ie a happy medium.  None
of them seem to be viable.  There are other functions besides
DisconnectEx() that do the same thing, so no benefit there;
examples are shutdown() and WSASendDisconnect().  Duplicating
a socket is an untested possibility, but I am doubtful that it
would prevent the DisconnectEx() from successfully shutting down
the connection.  Adding sleeps (and removing DisconnectEx()) to
allow closesocket() to always close gracefully might work but
introduces undesired delays in processing.  The socket SO_LINGER
option was tested because it is intended for this very thing...
delay closing the socket until sufficient time has passed to
make sure all sent messages have been received by the other side.
This option does not appear to work (ie, no impact) for asyncronous
(overlapped) sockets as used by the driver; Microsoft doc
specifically recommends against using SO_LINGER with overlapped
sockets anyway, though doesn't specifically state why.

Lastly, to improve the diagnostics, a check of the return code from
closesocket() in GCC_DISCONNECT was added with tracing.

------------------------------------------------------------------------
r3584 | loera01 | 2010-09-10 12:15:14 -0400 (Fri, 10 Sep 2010) | 18 lines

Created an new internal function, PutTinyInt(), which is similar to the
existing PutInt().  PutInt() copies an integer to the output data buffer.
PutTinyInt() is simlar to PutInt(), but used a value argument of an i1 instead
of an i2.  Apparently integers are aligned on 8-byte boundaries on Solaris
64-bit, so i2's were not reliable arguments.

Since 64-bit Solaris is a UCS4 machine, meaning that the SQLWCHAR variable
type is an i4 (wchar_t), particular care was required in converting to UCS2
from UCS4 and vice versa.  As each element of the string array was scanned
during the conversion, intermediate variables were required in order for each
string element to be aligned properly.  I actually had to use a union of two
u_i2's and one u_i4 when converting from UCS4 to UCS2 so that only the
lower-order two bytes of the UCS4 element were copied.

I had thrown in a call to GetWCharChar() at one point of the conversion, but
this had already been done previously.  This caused no error on Windows, but on Solaris, a garbage value resulted.  So I removed the reduncant function call.

------------------------------------------------------------------------