------------------------------------------------------------------------ r3584 | loera01 | 2010-09-10 12:15:14 -0400 (Fri, 10 Sep 2010) | 18 lines
Created an new internal function, PutTinyInt(), which is similar to the
existing PutInt().  PutInt() copies an integer to the output data buffer.
PutTinyInt() is simlar to PutInt(), but used a value argument of an i1 instead
of an i2.  Apparently integers are aligned on 8-byte boundaries on Solaris
64-bit, so i2's were not reliable arguments.

Since 64-bit Solaris is a UCS4 machine, meaning that the SQLWCHAR variable
type is an i4 (wchar_t), particular care was required in converting to UCS2
from UCS4 and vice versa.  As each element of the string array was scanned
during the conversion, intermediate variables were required in order for each
string element to be aligned properly.  I actually had to use a union of two
u_i2's and one u_i4 when converting from UCS4 to UCS2 so that only the
lower-order two bytes of the UCS4 element were copied.

I had thrown in a call to GetWCharChar() at one point of the conversion, but
this had already been done previously.  This caused no error on Windows, but on Solaris, a garbage value resulted.  So I removed the reduncant function call.

------------------------------------------------------------------------
r3585 | drivi01 | 2010-09-10 14:45:12 -0400 (Fri, 10 Sep 2010) | 8 lines

Added new help file gcnpwdvers.hlp to both 32-bit and 64-bit
installers.

Also added new parameter gcn.pwd_encode_version to the post
installer so it is installed on upgrades from earlier
releases.

------------------------------------------------------------------------