------------------------------------------------------------------------ r3592 | jenjo02 | 2010-09-13 16:30:17 -0400 (Mon, 13 Sep 2010) | 12 lines
In dm2d_open(), the database was being marked as not usable by MVCC if
it was opened read-only, in this issue due to the DB being marked
inconsistent, but would apply to consistent, read-only as well.

This cause dmx_begin() to not allocate a LG_CRIB for the transaction,
but dm2t_open() returned a lock level of MVCC to dmt_open() which
expected a CRIB to have been allocated.

There's no compelling reason MVCC should be disallowed in a read-only
DB, so remove that check from dm2d_open().

------------------------------------------------------------------------
r3593 | loera01 | 2010-09-13 18:30:15 -0400 (Mon, 13 Sep 2010) | 10 lines

In change 505621, I had thought the DBMS had changed its processing of
interval literals; as a result, I removed the 'interval' clause in the
converted query string.  Thus, "{ interval '5 12:04:56.123456' day to second }"
became "'5 12:04:56.123456' day to second".  It is now clear that the test used
for bug 122174 had faulty data.

I reinstated the code that preserved the "interval" clause, and odbc06 produced
no further diffs.

------------------------------------------------------------------------