------------------------------------------------------------------------ r3594 | hanal04 | 2010-09-14 04:45:17 -0400 (Tue, 14 Sep 2010) | 23 lines
Cross integration of change 507299

Review Board IP  http://reviewboard.ingres.prv/r/5775

In opc_sekeys() the outer key (seop->pst_left_pst_left) may be
a constant VLUP. The client test case induced this case by
having cache_dynamic enabled. If the outer key is a VLUP its
length will be ADE_LEN_UNKNOWN (-1). When the length is added
to the size of the DSH row required to hold the result we
decrement the buffer size and hit a buffer over-run during
the execution of the compiled SE join.
This change detects the VLUP outer key case and adds a
coercion to the CX. The VLUP outer key is coerced to the same
type as the inner subselect (seop->pst_left->pst_right).
This achieves two things:
1) We obtain and use the correct length adjustment for the
DSH result buffer. This sovles the reported SIGSEGV.
2) The non-variable length data types (such as BYTE) are padded
during the coercion so that the comparision of the outer key and
the inner subselect will return the correct number of rows.

------------------------------------------------------------------------
r3595 | abbjo03 | 2010-09-14 10:00:14 -0400 (Tue, 14 Sep 2010) | 12 lines

Cross-integrate change 507342 from ingres2006r3.
Changes to prevent hang due to unresolved distributed deadlock.

    In LK_distdeadlock_thread(), for shared lock lists,
    llb_stamp may not be the same value as lkb_lkreq.request_id.
    The message is stamped with lkb_lkreq.request_id, not
    llb_stamp, when originated; so check against the former.
    (this change had already been made in the main codeline).

    In CXdlm_get_blki, if a lock has been granted, update rqmode.

------------------------------------------------------------------------