------------------------------------------------------------------------ r3602 | abbjo03 | 2010-09-16 15:45:12 -0400 (Thu, 16 Sep 2010) | 3 lines
Changes needed to build 131 on main on VMS.

------------------------------------------------------------------------
r3603 | coomi01 | 2010-09-17 03:00:12 -0400 (Fri, 17 Sep 2010) | 33 lines

Cross integration of change 507395

Review Board IP  http://reviewboard.ingres.prv/r/5876

With two threads attempting to access values from the
same sequence, we may end up with a mutex deadlock
where the database (hence the sequence) is being
journaled.

The senario is that the advance thread may have got
to the point where it decides to replenish the sequence
cache. It will hold the SeqGen mutex at this stage.
It then proceeds down through dms_next_seq() to
procedure dm2d_check_db_backup(). Here it might
then attempt to take a DCB mutex.

Meanwhile the second thread could be validating some
other query that uses the sequence. qeq_validate()
holds the DCB mutext before examing sequences. That is
unfortunate, for in the call to FindSeq() it attempts
to get the SeqGen mutex.

The solution is to take ths DCB mutex in the advanced
routine whenever ther is a likelyhood that a call
to journalling may be made. The mutex is held until
the next sequence is fetched and journaled. Or, released
immediately if there is not need to advance the sequence.

If the rear guard thread has the mutex, the advanced
one will now have to wait a while. Likewise vice versa.

------------------------------------------------------------------------