------------------------------------------------------------------------ r3614 | smeke01 | 2010-09-22 07:00:09 -0400 (Wed, 22 Sep 2010) | 24 lines
Review Board IP  http://reviewboard.ingres.prv/r/5950

Cross integration of change 506313

In opn_combgen(), loosen heuristic 2 to allow the combination to include a placement index without its primary iff the index is joinable to at least one other table in the combination. Note that if this change proves with experience to be too lenient, a possible more restrictive alternative would be to allow the placement index to appear without the primary iff there are where-clause predicates that include a primary column and a column from at least one of the combination tables, and every such primary column appears in the placement index.

As a result of the change above, it is now possible for a placement index to come last in the permutation. So in opn_permgen(), remove the heuristic that a placement index cannot cannot appear last in a permutation. However, to be cautious, keep the heuristic that insists that where a placement index appears first or second, the corresponding primary table must appear next. Note that if this function is re-written and this heuristic is kept, it should be generalised to values of n > 3.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3615 | drivi01 | 2010-09-22 09:00:10 -0400 (Wed, 22 Sep 2010) | 10 lines

PROPAGATE change #507481 from ingres10 to main:

qp232 hangs on Windows x64. qp589 is producing a diff.

This change fixes optimization on dmse.c file which is
what's causing a hang in qp232.sep.  It also fixes
qp589.sep to use batch script instead of shell to
avoid portability issues such as seen on this platform.

------------------------------------------------------------------------