------------------------------------------------------------------------ r3615 | drivi01 | 2010-09-22 09:00:10 -0400 (Wed, 22 Sep 2010) | 10 lines
PROPAGATE change #507481 from ingres10 to main:

qp232 hangs on Windows x64. qp589 is producing a diff.

This change fixes optimization on dmse.c file which is
what's causing a hang in qp232.sep.  It also fixes
qp589.sep to use batch script instead of shell to
avoid portability issues such as seen on this platform.

------------------------------------------------------------------------
r3616 | hanje04 | 2010-09-22 18:45:10 -0400 (Wed, 22 Sep 2010) | 7 lines

Part deux of the fix for this bug. Part one fixed the GUI but not the response
file generation which was just messed up. Correct case statement for processing
install options in addMiscOps(). Also add new function to assist with this:

    addMiscOp()

------------------------------------------------------------------------