------------------------------------------------------------------------ r3619 | hanje04 | 2010-09-23 13:15:12 -0400 (Thu, 23 Sep 2010) | 12 lines
The GUI installer has never properly supported installation of the
ingres-32bit RPM package which provides the 32bit runtime environment for
64bit reverse hybrid builds. There is now also a bug in ingpkgqry.py that
causes the GUI to error when removing an existing installation if the
ingres-32bit RPM package is installed.

As a first step to supporting installation and removal of this package, add
it to the list of packages to be installed for a "full installation" and
correct the package name in ingpkgqry.py so that the 32bit package is found
and added to the list of packages to be removed along with the rest.

------------------------------------------------------------------------