------------------------------------------------------------------------ r3619 | hanje04 | 2010-09-23 13:15:12 -0400 (Thu, 23 Sep 2010) | 12 lines
The GUI installer has never properly supported installation of the
ingres-32bit RPM package which provides the 32bit runtime environment for
64bit reverse hybrid builds. There is now also a bug in ingpkgqry.py that
causes the GUI to error when removing an existing installation if the
ingres-32bit RPM package is installed.

As a first step to supporting installation and removal of this package, add
it to the list of packages to be installed for a "full installation" and
correct the package name in ingpkgqry.py so that the 32bit package is found
and added to the list of packages to be removed along with the rest.

------------------------------------------------------------------------
r3620 | huazh01 | 2010-09-23 14:30:11 -0400 (Thu, 23 Sep 2010) | 37 lines

Review Board IP  http://reviewboard.ingres.prv/r/5979

Cross integration of change 506603

Issue: 144064, 144218, 142217
Bug: 123615, 124187

b123615:

on psy_view(), we now search for nested derived table after adding
dtmap into viewmap. This prevents infinite loop in opa_pview().

b124187:

This is a side effect of the fix to b115096.

on opa_link(), child agg must be visible (link to) SAGG if child's
bylist appears in the main subquery as non-printing resdoms.
Otherwise, the non-printing resdom will appear to be in an eqcls
by itself, causing a cart-prod and duplicated rows.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------