------------------------------------------------------------------------ r3622 | horda03 | 2010-09-24 06:00:49 -0400 (Fri, 24 Sep 2010) | 36 lines
Review Board IP  http://reviewboard.ingres.prv/r/6001

Cross integration of change 507546

n VMS (and possibly other platforms) if the do loop
in dm2u_convert is exited before the DMF_TCB variable (t)
is assigned (easily reproduced by defining a location creating
a table that uses the location; stop ingres, change the physical
path to the directory; start Ingres and UPGRADEDB the DB -
the upgradedb will need to "upgrade" he table though).

When the do loop is exited, the next instruction examined the
contents of the "t" variable, if the variable had not been
assigned, then a SIGSEGV is likely.

dm2uconv.c

Check status and rcb before using the DMF_TCB pointer, if not OK
or RCB is null then t may not be defined.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3623 | kiria01 | 2010-09-24 08:30:16 -0400 (Fri, 24 Sep 2010) | 23 lines

Plug length check hole with parameter key indexing.

A previous change for parameter marks applied an explicit
coercion to the parameter if it could determine the expected
datatype and size from a corresponding VAR node across a
comparison operator. This had the effect of removing unwanted
conversions to Unicode when the datatype of the column being
indexed was not Unicode. However, it was possible for the
parameter to specify a key exceeding the column width that
when truncated might match data that the non-truncated form
would not have.

This change adds an additional boolean factor as follows:
  column = ?
became:
  column = coercion(?,len)
and now becomes:
  column = coercion(?,len) AND ? = coercion(?, len)
If the secondary compare fails then so would 'column = ?'
but 'column = coercion(?,len)' might make a match.

------------------------------------------------------------------------