------------------------------------------------------------------------ r3623 | kiria01 | 2010-09-24 08:30:16 -0400 (Fri, 24 Sep 2010) | 23 lines
Plug length check hole with parameter key indexing.

A previous change for parameter marks applied an explicit
coercion to the parameter if it could determine the expected
datatype and size from a corresponding VAR node across a
comparison operator. This had the effect of removing unwanted
conversions to Unicode when the datatype of the column being
indexed was not Unicode. However, it was possible for the
parameter to specify a key exceeding the column width that
when truncated might match data that the non-truncated form
would not have.

This change adds an additional boolean factor as follows:
  column = ?
became:
  column = coercion(?,len)
and now becomes:
  column = coercion(?,len) AND ? = coercion(?, len)
If the secondary compare fails then so would 'column = ?'
but 'column = coercion(?,len)' might make a match.

------------------------------------------------------------------------
r3624 | jenjo02 | 2010-09-24 15:15:16 -0400 (Fri, 24 Sep 2010) | 24 lines

Changes for B118915 causes pass-abort recovery in the RCP
to fail.

This happens because dmxe_abort() returns E_DB_OK even when
transaction recovery fails and is handed over to the RCP
for pass-abort recovery, then dmx_abort() releases the
failed transaction's lock list, which is still needed by
the RCP.

Now, when dmxe_abort(DMXE_DELAY_UNLOCK) performs pass-abort,
it returns E_DB_ERROR with the error E_DM9509_DMXE_PASS_ABORT.
On seeing this, dmx_abort() will not release the transaction's
lock list.

Also fixed a couple of other minor problems encountered while
solving the pass-abort problem:

Fixed dmfrecover to not perform an archive cycle during
online recovery.

Strengthened LKcreate_list(LK_RECOVER_LLID)
validation of lock list.

------------------------------------------------------------------------