------------------------------------------------------------------------ r3630 | horda03 | 2010-09-28 06:30:12 -0400 (Tue, 28 Sep 2010) | 23 lines
Review Board IP  http://reviewboard.ingres.prv/r/6037

Cross integration of change 507587

Thought this might be the cause of issue 145697, but alas
this code isn't in 9.1. BUt thought better fix anyhow.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3631 | hanje04 | 2010-09-28 13:00:11 -0400 (Tue, 28 Sep 2010) | 16 lines

The initial change to fix this bug revealed another problem in the installation
of the ingres-32bit package. Namely that the response file api (rfapi) isn't
aware of the 32bit package. When a "complete" installation is performed, an
error occurs during response file creation, which isn't properly reported in
the GUI (only in the tracing, will resolve in a separate IP). This results in
the installation be performed without a response file.

Address this by adding support for the ingres-32bit package to the rfapi and
correcting the info in the GUI.

Define II_COMPONENT_32BIT in rfapi.h
Add ii_component_32bit structure in rfapidata.c
Add ii_component_32bit to pkg_info in rfapidata.c
Correct rfapi_name in pkg_sup32 in gipdata.c

------------------------------------------------------------------------