------------------------------------------------------------------------ r3631 | hanje04 | 2010-09-28 13:00:11 -0400 (Tue, 28 Sep 2010) | 16 lines
The initial change to fix this bug revealed another problem in the installation
of the ingres-32bit package. Namely that the response file api (rfapi) isn't
aware of the 32bit package. When a "complete" installation is performed, an
error occurs during response file creation, which isn't properly reported in
the GUI (only in the tracing, will resolve in a separate IP). This results in
the installation be performed without a response file.

Address this by adding support for the ingres-32bit package to the rfapi and
correcting the info in the GUI.

Define II_COMPONENT_32BIT in rfapi.h
Add ii_component_32bit structure in rfapidata.c
Add ii_component_32bit to pkg_info in rfapidata.c
Correct rfapi_name in pkg_sup32 in gipdata.c

------------------------------------------------------------------------
r3632 | shust01 | 2010-09-29 14:15:14 -0400 (Wed, 29 Sep 2010) | 29 lines

Review Board IP  http://reviewboard.ingres.prv/r/6056

Cross integration of change 507597

When creating a response file (using setup /c) and a traditional
installation is chosen, that information does not appear in
the response file.  The other types of installs do add a line.

Change involved writing II_CONFIG_TYPE="TRAD" to the response
file when a traditional installation is chosen and checking
for 'TRAD' when the response file is read (instead of the word
'NULL' which doesn't match what is checked on Linux).

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------