------------------------------------------------------------------------ r3632 | shust01 | 2010-09-29 14:15:14 -0400 (Wed, 29 Sep 2010) | 29 lines
Review Board IP  http://reviewboard.ingres.prv/r/6056

Cross integration of change 507597

When creating a response file (using setup /c) and a traditional
installation is chosen, that information does not appear in
the response file.  The other types of installs do add a line.

Change involved writing II_CONFIG_TYPE="TRAD" to the response
file when a traditional installation is chosen and checking
for 'TRAD' when the response file is read (instead of the word
'NULL' which doesn't match what is checked on Linux).

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------