------------------------------------------------------------------------ r3641 | bonro01 | 2010-10-01 15:30:18 -0400 (Fri, 01 Oct 2010) | 7 lines
Fix Windows bldenv.bat script to use preset values for MKSLOC
and SHELL so that the script does not prompt for them.
Fix Jamrules for Windows to use correct slashes for filename separators.
And use the $(RM) command for file deletions because the cygwin rm command
does not understand windows backslashes.

------------------------------------------------------------------------
r3642 | loera01 | 2010-10-01 19:00:17 -0400 (Fri, 01 Oct 2010) | 10 lines

Additions to the tinyint.c test used for testing bug 124517 showed errors for
coercions of SQL_C_UTINYINT and SQL_C_DEFAULT to SQL_TINYINT.

In PutParm(), the tinyValue buffer used for the PutTinyInt() routine was
missing in the case of SQL_C_UTINYINT, so I added the reference to tinyValue.

The SQL_C_STINYINT case was mistakenly grouped with SQL_C_UTINYINT elsewhere
in PutParm(), so I grouped SQL_C_STINYINT with SQL_C_TINYINT.

------------------------------------------------------------------------