------------------------------------------------------------------------ r3649 | drivi01 | 2010-10-04 15:00:14 -0400 (Mon, 04 Oct 2010) | 16 lines
On Windows, replicator is throwing a segmentation violation
when altering CDDS from VDBA.  Under the covers, VDBA
calls ddgensup where the SEGV occurs.  The SEGV is due to
the hardcoded buffer in that utility.  After the SEGV was fixed
the altering of CDDS was completing with obscure errors.  After
some investigations, errors were caused by the column length.

It is documented in the Replicator guide that long id columns
can not exceed 252 characters.

This change fixes the SEGV by removing a hardcoded buffer and
also adds some column length checking before registering
columns in the replicator to make sure they do not exceed the
252 character length.

------------------------------------------------------------------------
r3650 | drivi01 | 2010-10-04 15:15:12 -0400 (Mon, 04 Oct 2010) | 12 lines

On Windows 2008, sepchild.exe hangs on Windows 2008 when the
privileges are dropped to execute ingres tasks at the user
level. This is something sepcific to Windows 2008 and requires
more investigation.

This change is a temporary fix to bypass the Vista routines
in sepchild and will execute all the commands at administrative
level.  This will eliminate the hang until the hang is resolved.

This change fixes this response file option.

------------------------------------------------------------------------