------------------------------------------------------------------------ r3650 | drivi01 | 2010-10-04 15:15:12 -0400 (Mon, 04 Oct 2010) | 12 lines
On Windows 2008, sepchild.exe hangs on Windows 2008 when the
privileges are dropped to execute ingres tasks at the user
level. This is something sepcific to Windows 2008 and requires
more investigation.

This change is a temporary fix to bypass the Vista routines
in sepchild and will execute all the commands at administrative
level.  This will eliminate the hang until the hang is resolved.

This change fixes this response file option.

------------------------------------------------------------------------