------------------------------------------------------------------------ r3653 | maspa05 | 2010-10-05 04:00:22 -0400 (Tue, 05 Oct 2010) | 23 lines
Review Board IP  http://reviewboard.ingres.prv/r/6099

Add some validation of SC930 log location.

Amend ult_set_tracefile() so that it checks the existence of the new location.
Previously it set it whatever the value, but on trying to use it, if we
failed to open the file we silently switched off tracing. ult_set_tracefile()
now returns a STATUS value so that we can tell whether it was successfully
set. If not we revert to the previous good value of the log location (if there
was one). Also add some TRdisplays so the user can see what has happened.

Exposed ult_tracefile_loc() so we can call it from outside ult. It now
returns a PTR so we don't have to include lo.h everywhere.

Changed pslsgram.yi so that we check whether ult_set_tracefile() failed and
if so then raise an error for the user. This will either be E_UL0400 or
E_UL0401. These are essentially the same text but one is when we revert to
an old value and the other when we revert to being unset.

Amended scf_trace so that we don't switch on tracing if the log directory
has not been set. Also output a message to this effect.

------------------------------------------------------------------------
r3654 | maspa05 | 2010-10-05 04:15:14 -0400 (Tue, 05 Oct 2010) | 14 lines

Review Board IP  http://reviewboard.ingres.prv/r/6119

SC930 support for batch execution.

Add PARM and QRY line output to pst_cpdata() which is the routine that
does the copy-insert optimization.

Amend scs_sequencer() so that current EQY output gets its rowcount from
the latest GCA response message in the list in the case of batch execution.

This fix does not affect the output format so the SC930 version has not been
incremented.

------------------------------------------------------------------------