------------------------------------------------------------------------ r3655 | smeke01 | 2010-10-05 05:45:11 -0400 (Tue, 05 Oct 2010) | 26 lines
Review Board IP  http://reviewboard.ingres.prv/r/6059

Cross integration of change 507575

For QUEL store away as a 'highwater mark' the value of the
pst_pmspec flag, so that we can use it if we do convert the operator into
one constant. For SQL, the pst_pmspec flag for the constant will always be
set to PST_PMNOTUSED.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3656 | schka01 | 2010-10-05 08:00:12 -0400 (Tue, 05 Oct 2010) | 13 lines

Eliminate the VARFI flag from the text -> ALL coercion,
which was preventing the coercion from being constant-folded.
Any coercions which might have runtime variances, such as
text -> decimal, have individual FI's.
The lack of constant folding is just a minor inefficiency
in standard Ingres;  in the Vectorwise code line it was a
problem, as the VW engine doesn't support the old quel types
C or TEXT.  In order to keep the code lines close, the
change is being propagated to main even though it's not a
big deal there.

------------------------------------------------------------------------