------------------------------------------------------------------------ r3659 | ashco01 | 2010-10-05 11:00:11 -0400 (Tue, 05 Oct 2010) | 9 lines
On VMS, the only valid command tail parameters to INGSTOP are -F and -FORCE.
If INGSTOP is called accidentally supplying a parameter valid for INGSTOP
on another OS (Unix or Windows), for example 'ingstop -rmcmd', then the
complete installation will be shutdown, which is not the action the
user intended.

Also removed redundant VMS OS Version check.

------------------------------------------------------------------------
r3660 | hanje04 | 2010-10-05 12:30:10 -0400 (Tue, 05 Oct 2010) | 6 lines

Move file.close() outside of while loop used to parse symbol.tbl in findInstID()
so that routine doesn't error if II_INSTALLATION insn't the first element
in sybmol.tbl (which is almost always is).

------------------------------------------------------------------------