------------------------------------------------------------------------ r3671 | hanal04 | 2010-10-08 03:45:12 -0400 (Fri, 08 Oct 2010) | 14 lines
Cross integration of change 507761

Review Board IP  http://reviewboard.ingres.prv/r/6166

Added a PSL_MDF_NO_OP to the modify operation table entry for
table_debug. This flag is subsequently used to set
DMU_NODEPENDENCY_CHECK to avoid the operation being blocked
by dependency checking. This new flag must be used with caution
as dependency checking is used to maintain data integrity.
PSL_MDF_NO_OP must only be used in operations that are
guaranteed not to break ANY dependencies.

------------------------------------------------------------------------
r3672 | hanal04 | 2010-10-08 04:00:12 -0400 (Fri, 08 Oct 2010) | 15 lines

Cross integration of change 507762

Review Board IP  http://reviewboard.ingres.prv/r/6152

Extend pss_stmt_flags to i8 from i4 to make room for a new flag
(PSS_STMT_NO_CACHEDYN) which can be used to disable the use
of cache dynamic on a per statement basis.
When VLUPs are used in the target list and cache dynamic is
enabled we hit E_AD2005 errors in coercions. This change
turns on PSS_STMT_NO_CACHEDYN if we process a VLUP whilsts
processing the target list. The flag is later used to
ensure repeat_dyn is not set to true for this class of query.

------------------------------------------------------------------------