------------------------------------------------------------------------ r3672 | hanal04 | 2010-10-08 04:00:12 -0400 (Fri, 08 Oct 2010) | 15 lines
Cross integration of change 507762

Review Board IP  http://reviewboard.ingres.prv/r/6152

Extend pss_stmt_flags to i8 from i4 to make room for a new flag
(PSS_STMT_NO_CACHEDYN) which can be used to disable the use
of cache dynamic on a per statement basis.
When VLUPs are used in the target list and cache dynamic is
enabled we hit E_AD2005 errors in coercions. This change
turns on PSS_STMT_NO_CACHEDYN if we process a VLUP whilsts
processing the target list. The flag is later used to
ensure repeat_dyn is not set to true for this class of query.

------------------------------------------------------------------------
r3673 | kiria01 | 2010-10-08 06:30:40 -0400 (Fri, 08 Oct 2010) | 11 lines

Add missing CTE support to ESQL

Documentation Impact:

None

Added the WITH element expressions to the ESQL grammer to cover
that supported in the main parser.

------------------------------------------------------------------------