------------------------------------------------------------------------ r3679 | loera01 | 2010-10-08 12:45:14 -0400 (Fri, 08 Oct 2010) | 14 lines
This change revises the CLI treatment of Unicode characters similar to the
changes made to the ODBC driver for 64-bit Solaris.  Conversions of each
character in ConvertUCS2ToUCS4() and ConvertUCS4ToUCS2() were done in three
steps instead of one, with I2ASSIGN_MACRO or I4ASSIGN_MACRO extracting each
string element initially, with an intermediate step assigning the UCS4 or UCS2
character.

I noticed another problem with SQLConnectW() and SQLDriverConnectW(). If an
error was found with ConvertWCharToChar(), SQL_ERROR was returned without
inserting any error information into the error block, resulting in a
segmentation violation.  I added the code to return the proper error
information.

------------------------------------------------------------------------