------------------------------------------------------------------------ r3681 | drivi01 | 2010-10-08 14:15:10 -0400 (Fri, 08 Oct 2010) | 17 lines
PROPAGATE change #507816 from ingres10 to main:

On Windows, during upgrade from any previous release that installed
demodb.  Post-installer fails to reload demodb data if user chose
not to upgrade user databases.

This is because demodb isn't handled in the post-installer like a
system database, instead it is treated like a user database.
Because of this, when demodb is being reloaded it fails because
the database was never upgraded.

This change adds demodb to the list of databases being upgraded
during configuration stage.  This will ensure the database
can be reloaded successfully.  demodb is a database installed
by Ingres and therefore should be maintained by Ingres.

------------------------------------------------------------------------
r3682 | miket | 2010-10-08 18:30:30 -0400 (Fri, 08 Oct 2010) | 8 lines

Trac Number: #584

Bug Number: 122403 124543

Description of changes:
MODIFY <table> ENCRYPT does not reconstruct the table, so do not worry
about dependency checking. Avoid it by setting PSL_MDF_NO_OP.

------------------------------------------------------------------------