------------------------------------------------------------------------ r3683 | stephenb | 2010-10-10 20:30:11 -0400 (Sun, 10 Oct 2010) | 16 lines
INTEGRATION: Change 507828 in ingres on 10-oct-2010 by stephenb on usrc-ingdevserver1_ingres10_steve-b132c

When record space is taken from the record buffer for the batch copy optimization,
the current location pointer, qeu_cptr, may move to a non bus-aligned location.
This buffer is also used for the QEF_INS_DATA structures that describe the records.
If this structure address does not start on a bus boundary some processors will generate
a bus error at runtime when assigning values to the structure members.
This change pads the size of the current reserved block using ALIGN_RESTRICT
so that it is always the next multiple of ALIGN_RESTRICT larger than the required
size, thus ensuring that qeu_cptr is always incremented to a bus boundary

Related Service Desk Issues:  147158

Build Considerations:  NONE

------------------------------------------------------------------------
r3684 | frima01 | 2010-10-11 01:00:29 -0400 (Mon, 11 Oct 2010) | 3 lines

Changed iisudbms to always upgrade demodb if existent.

------------------------------------------------------------------------