------------------------------------------------------------------------ r3693 | hanal04 | 2010-10-12 04:00:13 -0400 (Tue, 12 Oct 2010) | 16 lines
Cross integration of change 507832

Review Board IP  http://reviewboard.ingres.prv/r/6219

In psl_lst_elem() when the qmode is PSQ_CONS the dmu_cb is not
available. If compression settings were specified we SEGV'd
trying to reference the dmu_cb fields.
Store the compression settings in the pss_curcons->pss_restab
for later use. This avoids the SEGV but further changes were
needed to actually ensure the index was compressed as intended:
Update pslcons.c to add qtxt related to compression setting.
Update opcahd.c to pass on the compression setting to QEF.
Update qeaddl.c to add qtxt related to compression setting.

------------------------------------------------------------------------
r3694 | kiria01 | 2010-10-12 08:45:10 -0400 (Tue, 12 Oct 2010) | 18 lines

Add missing CTE support to ESQL

Documentation Impact:

None

Added the WITH element expressions to the ESQL grammer to cover
that supported in the main parser.

The DECLARE id CURSOR FOR production erroneously reduces to error
if the 'empty' prodution is involved when it should have accepted
the string literal. Re-writing the grammar for csr_select_sub to be
essentially two productions, one with query_with expanded and the
other with it omitted, avoid the problem. However, iyacc may have a
bug in this area.

------------------------------------------------------------------------