------------------------------------------------------------------------ r3694 | kiria01 | 2010-10-12 08:45:10 -0400 (Tue, 12 Oct 2010) | 18 lines
Add missing CTE support to ESQL

Documentation Impact:

None

Added the WITH element expressions to the ESQL grammer to cover
that supported in the main parser.

The DECLARE id CURSOR FOR production erroneously reduces to error
if the 'empty' prodution is involved when it should have accepted
the string literal. Re-writing the grammar for csr_select_sub to be
essentially two productions, one with query_with expanded and the
other with it omitted, avoid the problem. However, iyacc may have a
bug in this area.

------------------------------------------------------------------------